+++ _Data/Maps.lua
@@ -1,6 +1,13 @@
 local addonName, addon = ...
 
 addon.NoCutsceneSkipMaps = {
-    21,
-    1169
+    [21] = {
+        ["Conditions"] = {}
+    },
+    [371] = {
+        ["Conditions"] = {29827},
+    },
+    [1169] = {
+        ["Conditions"] = {}
+    }
 }